gcp/requirements.txt
pandas
google-cloud-bigquery
SQLAlchemy